+++ app/src/main/java/com/practicum/playlistmaker/utils/converters/DimensionConverter.kt
@@ -12,16 +12,3 @@ fun dimensionsFloatToIntConvert(dim: Float, context: Context): Int {
         context.resources.displayMetrics
     ).toInt()
 }
-
-fun isoDateToYearConvert(isoDate: String): String {
-    var year: String
-    val isoDateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'")
-    val targetDateFormat = SimpleDateFormat("yyyy")
-    try {
-        val date: Date = isoDateFormat.parse(isoDate)
-        year = targetDateFormat.format(date)
-    } catch (e: Exception) {
-        year = ""
-    }
-    return year
-}
