+++ app/src/main/java/com/practicum/playlistmaker/utils/converters/dateConvert.kt
@@ -1,0 +1,17 @@
+package com.practicum.playlistmaker.utils.converters
+
+import java.text.SimpleDateFormat
+import java.util.Date
+
+fun isoDateToYearConvert(isoDate: String): String {
+    var year: String
+    val isoDateFormat = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'")
+    val targetDateFormat = SimpleDateFormat("yyyy")
+    try {
+        val date: Date = isoDateFormat.parse(isoDate)
+        year = targetDateFormat.format(date)
+    } catch (e: Exception) {
+        year = ""
+    }
+    return year
+}
