+++ app/src/main/java/com/practicum/playlistmaker/utils/converters/imageResizeConvert.kt
@@ -1,0 +1,5 @@
+package com.practicum.playlistmaker.utils.converters
+
+fun getCoverArtwork(artworkUrl100: String): String {
+    return artworkUrl100.replaceAfterLast('/',"512x512bb.jpg")
+}
